Fix bitness-dependent IBI header output
When IBIze outputs the IBI header, it calculates the number of required
bytes by taking the size of a pointer. This differs between 32-bit (4
bytes) and 64-bit (8 bytes) systems, even though the file format itself
prescribes 4 bytes. Modify the code to take the length of the string
(including the terminating NUL byte) instead.