Make IBIze build on modern Linux and Mac OS systems (#5)
* Set correct subdirectory in CMakeLists.txt

Paths are case sensitive on Linux and Mac OS system, so specify the
correct subdirectory.

* Use C++11 standard for CMake

Use the C++11 standard when building with CMake, as e.g. C++17
introduced std::byte which is a conflicting type with the one declared
in Ibize.

* ibize: Add missing includes

Add any missing includes for Ibize to compile with a modern version of
GCC, in particular using the standard C++11 or newer.

Also use C++ like include headers instead of C style ones.